+++ capytaine/green_functions/libDelhommeau/src/constants.f90
@@ -3,6 +3,7 @@
 
 MODULE CONSTANTS
 
+  USE ieee_arithmetic
   USE FLOATING_POINT_PRECISION, ONLY: PRE
 
   IMPLICIT NONE
@@ -11,6 +12,8 @@ MODULE CONSTANTS
   REAL(KIND=PRE), PARAMETER :: ONE = 1
 
   REAL(KIND=PRE), PARAMETER :: PI = 3.141592653588979 ! π
+  REAL(KIND=PRE), PARAMETER :: EULER_GAMMA = 0.5772156649
+  REAL(KIND=PRE), PARAMETER :: LOG_2 = LOG(REAL(2d0, kind=pre))
   COMPLEX(KIND=PRE), PARAMETER :: II = (0, 1)         ! Imaginary unit
 
   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -26,5 +29,13 @@ MODULE CONSTANTS
   integer, parameter :: HIGH_FREQ = 0  ! legacy from Nemoh
   integer, parameter :: LOW_FREQ = 1  ! aka XieDelhommeau
 
+  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+contains
+
+  PURE LOGICAL FUNCTION is_infinity(x)
+    REAL(KIND=PRE), INTENT(IN) :: x
+    is_infinity = (.NOT. ieee_is_finite(x))
+  END FUNCTION
 
 END MODULE CONSTANTS